+++ docs/src/whats-new.md
@@ -2,6 +2,13 @@
 
 - [`loglayer` Changelog](/core-changelogs/loglayer-changelog)
 
+## Jan 18, 2025
+
+Add new options to the `ConsoleTransport`:
+
+- `messageField`: Allows you to specify the field in the log message object where the message should be stored. This is useful when you want to log structured data and need to specify the field name for the message.
+- `level`: Sets the minimum log level to process. Messages with a lower priority level will be ignored.
+
 ## Jan 17, 2025
 
 - Added the [Log File Rotation Transport](https://loglayer.dev/transports/log-file-rotation) to write logs to files and rotate them